+++ web/src/lib/components/user-settings-page/sidebar-appearance.svelte
